drm: bridge: Move exynos_drm_dsi into bridges
MIPI DSI IP used in Exynos and i.MX8MM are based Samsung MIPI DSIM
bridge.

In order to access the common bridge between platforms the ideal
way to push the driver into the bridge area.

So, this patch will move the exynos_drm_dsi.c driver into samsung-dsim.c
and do the necessary function name changes.

This would help to improve the bridge driver in future patches in order
to make use of different SoC families and platforms.

The bridge now support two implementations,

A. With component_ops and exynos specific code exclusively for exynos
   dsi drivers and it's legacy bindings.

B. Without componenet_ops for newly implemented bridges and its users
   like i.MX8MM.

The future plan is to fix the implementation A) by dropping component_ops
and fixing exynos specific code in order to make the bridge more mature
to use.

Updated MAINTAINERS file for this bridge with exynos drm maintainers along
with Andrzej as he is the original author.

Tomasz Figa has been not included in MAINTAINERS as he is not available
via samsung.com.

Signed-off-by: Jagan Teki <jagan@amarulasolutions.com>