btrfs: handle DIO read faults properly
Dylan reported a problem where he had an io_uring test that was returning
short DIO reads with ee5b46a353af ("btrfs: increase direct io read size
limit to 256 sectors") applied.  This turned out to be a red herring,
this simply increases the size of the reads we'll attempt to do in one
go.  The root of the problem is that because we're now trying to read in
more into our buffer, we're more likely to hit a page fault while trying
to read into the buffer.

Because we pass IOMAP_DIO_PARTIAL into iomap if we get an -EFAULT we'll
simply return 0, expecting that we'll do the fault and then try again.
However since we're only checking for a ret > 0 || ret == -EFAULT we
return a short read.  Fix this by checking for a 0 read from iomap as
well.  I've left the EFAULT case because it appears like we can still
get this in the case of a page fault from a direct read from an inline
extent.

Jens tested this patch with their testcase and it fixed the problem.

Reported-by: Dylan Yudaken <dylany@fb.com>
Tested-by: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Josef Bacik <josef@toxicpanda.com>
Signed-off-by: David Sterba <dsterba@suse.com>